+++ source/Python/CBCT_custom_algorithm_example_PyOpenCL.py
@@ -15,6 +15,9 @@ from omegatomo import proj
 from omegatomo.util import CTEFOVCorrection
 from omegatomo.reconstruction.prepass import linearizeData
 from omegatomo.util.powermethod import powerMethod
+from omegatomo.util.priors import RDP
+from omegatomo.util.priors import NLReg
+from omegatomo.util.priors import TV
 import matplotlib as plt
 
 options = proj.projectorClass()
@@ -114,9 +117,11 @@ options.offsetCorrection = False
 # Should have no effect at the moment
 options.enforcePositivity = True
 
-# Unused at the moment
+# Regularization parameter
 options.beta = .1
-options.NLMsigma = 2.00e-3
+# Filter parameter for non-local regularization
+options.NLMsigma = 4.00e-3
+# Edge preservation parameter for RDP and NLRDP
 options.RDP_gamma = 10.
 
 options.verbose = 1
@@ -197,6 +202,11 @@ for k in range(options.Niter):
         p[i] = pl
         g1 = g1 + dg
         g = g1 + (theta * options.subsets) * dg
+		# Regularized versions are enabled by uncommenting one regularizer and the update equation
+        # grad = RDP(d_f, options.Nx, options.Ny, options.Nz, options.RDP_gamma, options.beta, rType = 3)
+        # grad = NLReg(d_f, options.Nx, options.Ny, options.Nz, options.NLMsigma, options.beta, NLType = 3, useAdaptive=True, adaptiveConstant=5e-6, rType = 3)
+        # grad = TV(d_f, options.Nx, options.Ny, options.Nz, options.beta, rType = 3)
+        # d_f = d_f - tau * (g + grad)
         d_f = d_f - tau * g
         d_f = cl.array.if_positive(d_f, d_f, zeroAr)
         print('Sub-iteration ' + str(i))