+++ source/opencl/opencl_functions_orth3D.h
@@ -73,8 +73,19 @@ DEVICE bool orthogonalHelper3D(const int tempi, const int uu, const uint d_N2, c
 	, const int TOFid
 #endif
 #endif
+#if defined(SPECT) && defined(ATN)
+	, float jelppi
+#endif
 #if defined(MASKBP) && defined(BP)
+#ifdef USEIMAGES
+#ifdef MASKBP3D
+	, const int ii, IMAGE3D maskBP
+#else
 	, const int ii, IMAGE2D maskBP
+#endif
+#else
+, const int ii, const CLGLOBAL uchar* CLRESTRICT maskBP
+#endif
 #endif
 #ifdef SPECT
     , const float coneOfResponseStdCoeffA, const float coneOfResponseStdCoeffB, const float coneOfResponseStdCoeffC, const float2 crXY, const float s1, const float sZ, const float diff2, const float center1, const float centerZ
@@ -115,6 +126,9 @@ DEVICE bool orthogonalHelper3D(const int tempi, const int uu, const uint d_N2, c
 		return true;
 	}
 #endif
+#if defined(SPECT) && defined(ATN)
+	local_ele *= jelppi;
+#endif
 #ifdef BP
 #if defined(MASKBP)
 	if (ii == 0) {
@@ -133,7 +147,11 @@ DEVICE bool orthogonalHelper3D(const int tempi, const int uu, const uint d_N2, c
 #endif
 #endif
 #else
+#ifdef MASKBP3D
+		const int maskVal = maskBP[tempi * d_N2 + uu * d_N3 + zz * d_Nxy];
+#else
 		const int maskVal = maskBP[tempi * d_N2 + uu * d_N3];
+#endif
 #endif
 		if (maskVal == 0)
 			return false;
@@ -190,8 +208,19 @@ DEVICE int orthDistance3D(const int tempi, const float diff1, const float diff2,
 	, const int TOFid
 #endif
 #endif
+#if defined(SPECT) && defined(ATN)
+	, float jelppi
+#endif
 #if defined(MASKBP) && defined(BP)
+#ifdef USEIMAGES
+#ifdef MASKBP3D
+	, const int ii, IMAGE3D maskBP
+#else
 	, const int ii, IMAGE2D maskBP
+#endif
+#else
+	, const int ii, const CLGLOBAL uchar* CLRESTRICT maskBP
+#endif
 #endif
 #ifdef SPECT
     , const float coneOfResponseStdCoeffA, const float coneOfResponseStdCoeffB, const float coneOfResponseStdCoeffC, const float2 crXY
@@ -241,6 +270,9 @@ DEVICE int orthDistance3D(const int tempi, const float diff1, const float diff2,
 				, TOFid
 #endif
 #endif
+#if defined(SPECT) && defined(ATN)
+				, jelppi
+#endif
 #if defined(MASKBP) && defined(BP)
 				, ii, maskBP
 #endif
@@ -270,6 +302,9 @@ DEVICE int orthDistance3D(const int tempi, const float diff1, const float diff2,
 #ifdef LISTMODE
 				, TOFid
 #endif
+#endif
+#if defined(SPECT) && defined(ATN)
+				, jelppi
 #endif
 #if defined(MASKBP) && defined(BP)
 				, ii, maskBP
@@ -314,6 +349,9 @@ DEVICE int orthDistance3D(const int tempi, const float diff1, const float diff2,
 				, TOFid
 #endif
 #endif
+#if defined(SPECT) && defined(ATN)
+				, jelppi
+#endif
 #if defined(MASKBP) && defined(BP)
 				, ii, maskBP
 #endif
@@ -343,6 +381,9 @@ DEVICE int orthDistance3D(const int tempi, const float diff1, const float diff2,
 #ifdef LISTMODE
 				, TOFid
 #endif
+#endif
+#if defined(SPECT) && defined(ATN)
+				, jelppi
 #endif
 #if defined(MASKBP) && defined(BP)
 				, ii, maskBP